ai-service/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.2
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
scikit-learn>=1.4.0
pandas>=2.1.4
numpy>=1.26.2
joblib>=1.3.2
redis>=5.0.1
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
httpx>=0.25.2
python-multipart>=0.0.6
scipy>=1.11.4
beautifulsoup4>=4.12.2
requests>=2.31.0
lxml>=4.9.3

